Purchases

+++ AutoArsenal-App/AutoArsenal-App/Controllers/PurchaseDetailsController.cs
@@ -111,7 +111,7 @@ namespace AutoArsenal_App.Controllers
                                     Quantity = reader.GetInt32(reader.GetOrdinal("Quantity")),
                                     ProductCategoryID = reader.GetInt32(reader.GetOrdinal("ProductCategoryId")),
                                     ManufacturerID = reader.GetInt32(reader.GetOrdinal("ManufacturerId")),
-                                    UnitPrice = reader.GetFloat(reader.GetOrdinal("UnitPrice"))
+                                    UnitPrice = reader.GetDouble(reader.GetOrdinal("UnitPrice"))
                                 };
                                 purchaseDetails.Add(item);
                             }
